+++ extensions/ql-vscode/src/language-support/contextual/cached-operation.ts
